refactor: split seq u32 → seq u16 + pkt_id u16 (zero wire overhead)
Wire bytes [17-20] split: seq (u16 LE, [17-18]) + pkt_id (u16 LE, [19-20]).
Same 23-byte header, same byte positions for all subsequent fields.

MOTIVATION
  A single seq field was serving three incompatible roles:
  - Application sequence (message ordering, Lamport hint)
  - Forwarding identity  (dedupe key (src_id, seq))
  - Control-plane token  (ROUTE_REQ / ROUTE_RPL correlation)
  Overloading one field made fallback floods impossible to implement
  correctly: preserving seq caused dedupe-drop at nodes that saw the
  original unicast; changing seq broke application-level ordering.

BEHAVIORAL CHANGES
  Fallback flood: fb.seq = pkt.seq (same logical message),
                  fb.pkt_id = ++g_ctrl_seq (fresh identity)
  → nodes that already saw the original unicast now forward the fallback.
  Retransmit (same pkt_id): correctly deduped — redundant delivery suppressed.
  Jitter seeding: uses pkt_id → fallback floods spread independently.
  Control-plane: seq = correlation token, pkt_id = unique injection ID.

FILES CHANGED
  firmware_core/protocol.h: struct seq u32 → seq u16 + pkt_id u16;
    PKT_ID_BYTE_OFFSET 19; wire layout comment updated.
  firmware_core/protocol.c: encode/decode split for both fields.
  firmware_core/routing.h:  PACKET IDENTITY MODEL block added; all
    signatures using seq → pkt_id; routing_build_route_{req,rpl} gains
    pkt_id param.
  firmware_core/routing.c:  dedupe keys on pkt_id; jitter seed uses
    pkt_id; routing_build_route_{req,rpl} sets hdr.pkt_id.
  rivr_layer/rivr_sources.c: logs show seq=%u pkt_id=%u; Lamport uses
    seq; jitter uses pkt_id; ROUTE_RPL build updated.
  rivr_layer/rivr_sinks.c:  ROUTE_REQ build; fallback flood pkt_id bump;
    beacon seq+pkt_id from g_ctrl_seq.
  rivr_layer/rivr_cli.c:    CLI chat origination seq+pkt_id.
  firmware_core/main.c:     sim rounds R1-R8 use .pkt_id; R5/R6 comments
    updated from (src,seq) to (src,pkt_id).
  tests/test_acceptance.c:  all struct initializers gain .pkt_id; run1-7
    callers updated; new run8_pkt_identity() (11 checks). 171 PASS/0 FAIL.
  tests/replay.c:           pkt_id JSON field parsed; backward-compat
    fallback pkt_id=seq for traces without pkt_id key.
  tests/fuzz_ffi_harness.c: seq_seed cast to uint16_t for jitter call.
  tools/rivr_decode.c:      pkt_id added to JSON output after seq.
  docs/en/architecture.md:  Phase 11 section; Phase 10 dedupe/log refs
    updated from seq to pkt_id.

  171 PASS / 0 FAIL  (was 160 PASS / 0 FAIL)

+++ tests/fuzz_ffi_harness.c
@@ -141,7 +141,7 @@ int main(void)
         uint16_t max_j  = (len >= 10) ? (uint16_t)((buf[8] << 8) | buf[9]) : 200u;
         if (max_j == 0) max_j = 1;   /* avoid degenerate case */
 
-        uint16_t jitter = routing_jitter_ticks(src_seed, seq_seed, max_j);
+        uint16_t jitter = routing_jitter_ticks(src_seed, (uint16_t)seq_seed, max_j);
         assert(jitter <= max_j);
     }
